Bug Fixes and new commands
Made a few bugs fixes:

- If there are two inputs and one is closed the library will now detect that from the tally
- Updated XML model for Audio so that meter levels are doubles
- Updated XML model for Audio so that the meter levels are doubles

Added functionality
- Inputs are now comparable
- Can tell if an input is still a valid input
- Added more events to the input listener, inputRemove and dataChanged\
- Overlays and Transitions models are now comparable

Added new commands
- Remove Input
- Add Input

+++ src/main/java/uk/co/flakeynetworks/vmix/status/Audio.java
@@ -13,11 +13,11 @@ public class Audio {
     private boolean muted;
 
     @Attribute
-    private int meterF1;
+    private double meterF1;
 
     @Attribute
-    private int meterF2;
+    private double meterF2;
 
     @Attribute
-    private int headphonesVolume;
+    private double headphonesVolume;
 } // end of Audio
